# Add auto_bridge to CachingOptimizer