--- a/docs/src/reference/models.md
+++ b/docs/src/reference/models.md
@@ -37,6 +37,7 @@ copy_to
 ```@docs
 AbstractModelAttribute
 Name
+CoefficientType
 ObjectiveFunction
 ObjectiveFunctionType
 ObjectiveSense