--- a/src/Utilities/mockoptimizer.jl
+++ b/src/Utilities/mockoptimizer.jl
@@ -335,6 +335,10 @@ function MOI.get(mock::MockOptimizer, attr::MOI.AbstractModelAttribute)
     end
 end
 
+function MOI.get(mock::MockOptimizer, attr::MOI.CoefficientType)
+    return MOI.get(mock.inner_model, attr)
+end
+
 #####
 ##### Names
 #####