## What problem is this trying to solve

Consider this model in Clp
```Julia
model = Model(Clp.Optimizer)
@variable(model, x[1:2] >= 0)
optimize!(model)
```
**Pop quiz: how many caches are there?**
```Julia
julia> backend(model)
MOIU.CachingOptimizer{MOI.AbstractOptimizer,MOIU.UniversalFallback{MOIU.Model{Float64}}}
in state ATTACHED_OPTIMIZER
in mode AUTOMATIC
with model cache MOIU.UniversalFallback{MOIU.Model{Float64}}
  fallback for MOIU.Model{Float64}
with optimizer MOIB.LazyBridgeOptimizer{MOIU.CachingOptimizer{Clp.Optimizer,MOIU.UniversalFallback{MOIU.Model{Float64}}}}
  with 0 variable bridges
  with 0 constraint bridges
  with 0 objective bridges
  with inner model MOIU.CachingOptimizer{Clp.Optimizer,MOIU.UniversalFallback{MOIU.Model{Float64}}}
    in state ATTACHED_OPTIMIZER
    in mode AUTOMATIC
    with model cache MOIU.UniversalFallback{MOIU.Model{Float64}}
      fallback for MOIU.Model{Float64}
    with optimizer Clp.Optimizer
```
There are two! There is the outer cache, then a bridging layer, and then the inner cache. 
But there are `0` bridges added to this problem! So our two caches are just copies of each other.

Why do we need two? Because a user might write this
```Julia
model = Model(Clp.Optimizer)
@variable(model, x[1:2])
@constraint(model, x in MOI.Nonnegatives(2))
optimize!(model)
Julia> backend(model)
MOIU.CachingOptimizer{MOI.AbstractOptimizer,MOIU.UniversalFallback{MOIU.Model{Float64}}}
in state ATTACHED_OPTIMIZER
in mode AUTOMATIC
with model cache MOIU.UniversalFallback{MOIU.Model{Float64}}
  fallback for MOIU.Model{Float64}
with optimizer MOIB.LazyBridgeOptimizer{MOIU.CachingOptimizer{Clp.Optimizer,MOIU.UniversalFallback{MOIU.Model{Float64}}}}
  with 0 variable bridges
  with 1 constraint bridge
  with 0 objective bridges
  with inner model MOIU.CachingOptimizer{Clp.Optimizer,MOIU.UniversalFallback{MOIU.Model{Float64}}}
    in state ATTACHED_OPTIMIZER
    in mode AUTOMATIC
    with model cache MOIU.UniversalFallback{MOIU.Model{Float64}}
      fallback for MOIU.Model{Float64}
    with optimizer Clp.Optimizer
```

For the majority of users just trying to solve LPs and MIPs, this means that their first solve involves a double copy, and lots of extra inference sorting out the bridging functions.

## The alternative

If you knew there was going to be no bridging, you could write
```Julia
model = Model(Clp.Optimizer; bridge_constraints = false)
@variable(model, x[1:2] >= 0)
optimize!(model)
julia> backend(model)
MOIU.CachingOptimizer{MOI.AbstractOptimizer,MOIU.UniversalFallback{MOIU.Model{Float64}}}
in state ATTACHED_OPTIMIZER
in mode AUTOMATIC
with model cache MOIU.UniversalFallback{MOIU.Model{Float64}}
  fallback for MOIU.Model{Float64}
with optimizer Clp.Optimizer
```
and if you get it wrong, you get a nice error message:
```Julia
model = Model(Clp.Optimizer; bridge_constraints = false)
@variable(model, x[1:2])
julia> @constraint(model, x in MOI.Nonnegatives(2))
ERROR: Constraints of type MathOptInterface.VectorOfVariables-in-MathOptInterface.Nonnegatives are not supported by the solver, try using `bridge_constraints=true` in the `JuMP.Model` constructor if you believe the constraint can be reformulated to constraints supported by the solver.
```

Thus, one option is to change the default in JuMP from `bridge_constraints = true` to `bridge_constraints = false`.

Pros: faster. Users are more aware when bridges are used. Most users don't use bridges
Cons: breaking. But it's a one-line change for users.

## Proposed approach

Start with the equivalent of `bridge_constraints = false`. If, when adding a constraint, the constraint is unsupported, add a `full_bridge_optimizer`.

Pros: Opt-in at MOI level. Better performance without breaking at JuMP level.
Cons: More complexity. Might be good for users to see when bridges are used.

See the companion PR in JuMP: https://github.com/jump-dev/JuMP.jl/pull/2513 which contains benchmarks proving the efficacy.

## TODOs

- [ ] Decide whether to do this, or change JuMP's default.
- [x] Tests
- [x] Benchmarks
- [ ] Docs
- [x] A way to get the `Number` type, rather than hard-coding `Float64`
- [ ] A way for JuMP to add extra bridges?
- [x] Bikeshed the argument name

Closes #1156 
Closes #1249
Closes #1251