@blegat 2021-02-26
Solvers that benefit from this second case are the solvers that have an efficient `copy_to`. These solvers usually have a `copy_to` function that start by converting the model into a matrix form and then load it into the solver.
So here we really have 3 copies of the model:
User model -> Bridged model -> Matrix form
With your suggestion, it would be reduced to two:
User model -> Matrix form
So another way to do this would be to use the Matrix form for the Bridged model. That is, instead of using a `MOIU.Model` in the `CachingOptimizer`, we use a matrix format instead so that when we call `MOI.copy_to(::Optimtizer, ::MatrixForm)`, the solvers sees that it's already exactly the matrix form it needs, get the pointers of the vectors describing the sparse matrix and give that to the solver directly, no copy needed.