@odow 2021-02-28
The matrix form is only used for loading though. It gets discarded and gc'd after `copy_to`.

There could be a way for the optimizer to say here is my desired cache. The default could be
```Julia
function desired_cache(model::AbstractOptimizer)
    return Utilities.UniversalFallback(Utilities.Model{Float64}())
end
```
and matrix solvers could specify something else.