@blegat 2021-03-02
One issue with the approach in the PR is that the bridge layer is added when the user calls `supports_...`. However, we should only add it when the user actually calls `add_...`. To make `supports_...` work, we should indeed check if it is supported if a bridge layer is added but we should not modify `model.optimizer`.

Another concern is that it fixes the use case of JuMP which has a cache before the bridge layer but for other use cases such as meta-solvers, you want to do basically
```julia
optimizer = MOI.instantiate(solver, with_bridge_type=Float64)
MOI.copy_to(optimizer, src)
```
In this case, we don't get any benefit from this PR.
To solve both use cases and avoid the issue with `supports` discussed above we could do the following:
1) Replace
https://github.com/jump-dev/MathOptInterface.jl/blob/9552c68316ddab8d89736fc84ca0ff20a28bdda2/src/Bridges/bridge_optimizer.jl#L323-L325
by
```julia
function MOI.copy_to(dest::AbstractBridgeOptimizer, src::MOI.ModelLike; kws...)
    if # Nothing is going to be bridged anyway
        MOI.copy_to(dest.model, src)
    else
        MOIU.default_copy_to(dest, , src; kws...)
    end
end
```
2) Add an option for the `CachingOptimizer` and do
```julia
function MOI.copy_to(dest::CachingOptimizer, ...)
    if state != NO_OPTIMIZER && # the option is true
        MOI.copy_to(dest.optimizer, ...)
    else
        # Same than what's currently done
    end
end
```
So with this option, we avoid using the cache in case `copy_to` is called.
In case modifications are done that are not supported by the solver, we simple do not catch the error.
JuMP could safely use it as there is a second layer of cache on top that would catch it and empty the optimizer (containing the bridges, the (unused) second cache and optimizer).