@odow 2021-03-02
The reason to add bridges on a `supports` call is two-fold:
  1. If you're asking `supports`, we don't want to return `false` when actually calling `add_constraint` will work.
  2. We don't want to initialize a bridge every time we call `supports`, only to reinitialize it when we call `add_constraint`.  

I'll have a play with your suggestion. But it seems like there are a lot of edge cases with modification and deletion.