@blegat 2021-03-03
> The reason to add bridges on a `supports` call is two-fold:
>  1. If you're asking `supports`, we don't want to return `false` when actually calling `add_constraint` will work.
>  2. We don't want to initialize a bridge every time we call `supports`, only to reinitialize it when we call `add_constraint`.

Yes, for this reason, the only sensible way to make `supports` not have any effect is to have two fields `optimizer` and `bridged_optimizer` and a `Bool` indicating which one is active.
But that's going away from having MOI layers following the "Do one thing and do it well" principle.