@odow 2021-03-04
More JuMP benchmarks using the script below:
```Julia
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. example_diet.jl clp       
 17.319493 seconds (57.13 M allocations: 2.858 GiB, 8.30% gc time)
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. example_diet.jl clp --auto
  6.071365 seconds (18.02 M allocations: 925.261 MiB, 5.73% gc time)
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. example_diet.jl glpk      
 13.341120 seconds (44.97 M allocations: 2.257 GiB, 8.24% gc time)
(base) oscar@Oscars-MBP JuMP % ~/julia --project=. example_diet.jl glpk --auto
  8.593967 seconds (28.04 M allocations: 1.406 GiB, 8.13% gc time)
```
2 or 3x improvement in time-to-first-solve for models users care about. And this is not counting additional speedups we will get from precompilation.

## Code

```Julia
using JuMP, GLPK, Clp
function example_diet(optimizer, auto_bridge)
    categories = ["calories", "protein", "fat", "sodium"]
    category_data = Containers.DenseAxisArray([
        1800 2200;
        91   Inf;
        0    65;
        0    1779
        ], categories, ["min", "max"]
    )
    foods = [
        "hamburger", "chicken", "hot dog", "fries", "macaroni", "pizza",
        "salad", "milk", "ice cream",
    ]
    cost = Containers.DenseAxisArray(
        [2.49, 2.89, 1.50, 1.89, 2.09, 1.99, 2.49, 0.89, 1.59],
        foods
    )
    food_data = Containers.DenseAxisArray(
        [
            410 24 26 730;
            420 32 10 1190;
            560 20 32 1800;
            380  4 19 270;
            320 12 10 930;
            320 15 12 820;
            320 31 12 1230;
            100  8 2.5 125;
            330  8 10 180
        ], foods, categories
    )
    model = Model(optimizer, auto_bridge = auto_bridge)
    set_silent(model)
    @variables(model, begin
        category_data[c, "min"] <= nutrition[c = categories] <= category_data[c, "max"]
        buy[foods] >= 0
    end)
    @objective(model, Min, sum(cost[f] * buy[f] for f in foods))
    @constraint(model, [c in categories],
        sum(food_data[f, c] * buy[f] for f in foods) == nutrition[c]
    )
    optimize!(model)
    term_status = termination_status(model)
    @assert term_status == MOI.OPTIMAL
    @constraint(model, buy["milk"] + buy["ice cream"] <= 6)
    optimize!(model)
    @assert termination_status(model) == MOI.INFEASIBLE
    return
end

if length(ARGS) > 0
    auto = get(ARGS, 2, "") == "--auto"
    if ARGS[1] == "clp"
        @time example_diet(Clp.Optimizer, auto)
    else
        @assert ARGS[1] == "glpk"
        @time example_diet(GLPK.Optimizer, auto)
    end
end
```