@odow 2021-03-05
Talked to @blegat offline. The alternative is to add a `DIRECT` state to `CachingOptimizer` which just ignores the cache, and throws errors if things are not supported.