@mlubin 2021-03-06
This PR claims to close #1156. #1156 says:
> Rather than making an ad-hoc change, we should take the time to throughly document the caching optimizer system, choose a design that makes the most sense, and then implement that.

Where's the thorough documentation? :)

Changing the external state of the `CachineOptimizer` in `supports_*` is a bad code smell. Even changing the internal state would be strange because it's implied that `supports_*` has no side effects.

Could you elaborate on what the `DIRECT` state would look like?