@odow 2021-03-08
Latest timings on Julia 1.5.3
```julia
(base) oscar@Oscars-MBP scripts % ~/julia --project=. example_diet.jl clp 
 13.592436 seconds (35.24 M allocations: 1.775 GiB, 6.56% gc time)
(base) oscar@Oscars-MBP scripts % ~/julia --project=. example_diet.jl clp --auto
  5.070573 seconds (11.74 M allocations: 606.745 MiB, 4.71% gc time)
(base) oscar@Oscars-MBP scripts % ~/julia --project=. example_diet.jl glpk      
 10.806421 seconds (29.22 M allocations: 1.471 GiB, 7.37% gc time)
(base) oscar@Oscars-MBP scripts % ~/julia --project=. example_diet.jl glpk --auto
  5.876158 seconds (12.85 M allocations: 666.154 MiB, 3.98% gc time)
```

And 1.6-RC1
```Julia
(base) oscar@Oscars-MBP scripts % '/Applications/Julia-1.6.app/Contents/Resources/julia/bin/julia' --project=. example_diet.jl clp
 12.588525 seconds (32.80 M allocations: 1.842 GiB, 6.04% gc time, 33.69% compilation time)
(base) oscar@Oscars-MBP scripts % '/Applications/Julia-1.6.app/Contents/Resources/julia/bin/julia' --project=. example_diet.jl clp --auto
  5.087086 seconds (10.93 M allocations: 647.489 MiB, 3.38% gc time, 92.28% compilation time)
(base) oscar@Oscars-MBP scripts % '/Applications/Julia-1.6.app/Contents/Resources/julia/bin/julia' --project=. example_diet.jl glpk     
  9.666504 seconds (27.98 M allocations: 1.566 GiB, 5.27% gc time, 38.85% compilation time)
(base) oscar@Oscars-MBP scripts % '/Applications/Julia-1.6.app/Contents/Resources/julia/bin/julia' --project=. example_diet.jl glpk --auto
  6.029965 seconds (12.95 M allocations: 772.171 MiB, 3.82% gc time, 99.95% compilation time)
```

Still lots of room to improve, but we've gone from 17s to 5s on Clp, and 13s to 6s on GLPK.